--- frontend/pages/hubs/[hubUrl]/[subHub]/organisations.tsx
@@ -1,2 +1,0 @@
-export { getServerSideProps } from "../organisations";
-export { default } from "../organisations";
