+++ frontend/pages/hubs/[hubUrl]/[subHub]/organizations.tsx
@@ -1,0 +1,2 @@
+export { getServerSideProps } from "../organizations";
+export { default } from "../organizations";
