--- frontend/pages/hubs/[hubUrl]/[subHub]/projects.tsx
@@ -1,2 +1,0 @@
-export { getServerSideProps } from "../projects";
-export { default } from "../projects";
