+++ frontend/pages/hubs/[hubUrl]/browse.tsx
@@ -1,37 +1,72 @@
+import { GetServerSideProps } from "next";
+import React, { useContext } from "react";
 import { useRouter } from "next/router";
-import React, { useEffect, useRef } from "react";
-import Head from "next/head";
-import LoadingContainer from "../../../src/components/general/LoadingContainer";
+import { getHubBrowseTypeServerSideProps } from "../../../src/components/hub/getHubBrowseTypeServerSideProps";
+import HubPageLayout from "../../../src/components/hub/HubPageLayout";
+import BrowseProjectsContent from "../../../src/components/browse/BrowseProjectsContent";
+import { FilterProvider } from "../../../src/components/provider/FilterProvider";
+import UserContext from "../../../src/components/context/UserContext";
+import Cookies from "universal-cookie";
 
-export async function getServerSideProps() {
-  return { props: {} };
-}
+export const getServerSideProps: GetServerSideProps = (ctx) =>
+  getHubBrowseTypeServerSideProps(ctx, "projects");
 
-export default function HubBrowseRedirect() {
+const TYPES = ["projects", "organizations", "members"];
+
+export default function HubProjectsPage({
+  hubUrl,
+  subHubSegment,
+  filterChoices,
+  initialLocationFilter,
+  hubData,
+  hubThemeData,
+  linkedHubs,
+  allHubs,
+  isLocationHub,
+}: any) {
+  const cookies = new Cookies();
+  const token = cookies.get("auth_token");
+  const { locale } = useContext(UserContext);
   const router = useRouter();
-  const { hubUrl } = router.query;
-  const redirectedRef = useRef(false);
-  const localePrefix = router.locale && router.locale !== "en" ? `/${router.locale}` : "";
+  const browsePath = subHubSegment ? `/hubs/${hubUrl}/${subHubSegment}` : `/hubs/${hubUrl}`;
 
-  useEffect(() => {
-    if (redirectedRef.current || !hubUrl) return;
-    const hash = window.location.hash.replace("#", "");
-    let target = `/hubs/${hubUrl}/projects`;
-    if (hash === "organizations") target = `/hubs/${hubUrl}/organisations`;
-    else if (hash === "members") target = `/hubs/${hubUrl}/members`;
-    const search = window.location.search;
-    redirectedRef.current = true;
-    // Use window.location.replace instead of router.replace so the browser cancels
-    // any pending meta-refresh timer (router.replace is SPA navigation, which leaves the timer active)
-    window.location.replace(`${target}${search}`);
-  }, [router, hubUrl]);
+  const handleTabChange = (_event: React.SyntheticEvent, newValue: number) => {
+    const tab = TYPES[newValue];
+    const targetPath =
+      tab === "projects"
+        ? `${browsePath}/browse`
+        : tab === "organizations"
+        ? `${browsePath}/organizations`
+        : `${browsePath}/members`;
+    const params = new URLSearchParams(window.location.search);
+    router.push(`${targetPath}${params.toString() ? `?${params}` : ""}`);
+  };
 
   return (
-    <>
-      <Head>
-        <meta httpEquiv="refresh" content={`1;url=${localePrefix}/hubs/${hubUrl || ""}/projects`} />
-      </Head>
-      <LoadingContainer headerHeight={113} footerHeight={80} />
-    </>
+    <HubPageLayout
+      activeTab={0}
+      TYPES_BY_TAB_VALUE={TYPES}
+      handleTabChange={handleTabChange}
+      hubUrl={hubUrl}
+      subHubSegment={subHubSegment}
+      linkedHubs={linkedHubs}
+      hubData={hubData}
+      hubThemeData={hubThemeData}
+      allHubs={allHubs}
+      isLocationHub={isLocationHub}
+    >
+      <FilterProvider
+        filterChoices={filterChoices}
+        initialLocationFilter={initialLocationFilter}
+        locale={locale}
+        token={token}
+        hubUrl={hubUrl}
+      >
+        <BrowseProjectsContent
+          filterChoices={filterChoices}
+          initialLocationFilter={initialLocationFilter}
+        />
+      </FilterProvider>
+    </HubPageLayout>
   );
 }
